src/services/governance/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1
jsonschema>=4.20.0
pytest>=7.0.0
eval_type_backport>=0.2.0

